Minor fixes and some comments

+++ src/main/java/com/justbuyit/eventprocessor/subscription/CancelSubscriptionEventProcessor.java
@@ -1,21 +1,20 @@
 package com.justbuyit.eventprocessor.subscription;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.justbuyit.auth.ConnectionSigner;
 import com.justbuyit.dao.CompanyDAO;
 import com.justbuyit.entity.Company;
 import com.justbuyit.eventprocessor.EventProcessor;
+import com.justbuyit.exception.AccountNotFoundException;
 import com.justbuyit.exception.JustBuyItException;
 import com.justbuyit.model.event.subscription.CancelSubscriptionEvent;
 import com.justbuyit.model.event.subscription.CancelSubscriptionEvent.CancelSubscriptionPayload;
 import com.justbuyit.model.result.Result;
 
+/**
+ * Event processor for subsciption cancel events.
+ */
 public class CancelSubscriptionEventProcessor extends EventProcessor<CancelSubscriptionEvent> {
 
-    private final static Logger LOG = LoggerFactory.getLogger(CancelSubscriptionEventProcessor.class);
-    
     private CompanyDAO companyDAO;
 
     public CancelSubscriptionEventProcessor(ConnectionSigner connectionSigner, CompanyDAO companyDAO) {
@@ -25,12 +24,14 @@ public class CancelSubscriptionEventProcessor extends EventProcessor<CancelSubsc
 
     @Override
     protected Result processEvent(CancelSubscriptionEvent event) throws JustBuyItException {
-        LOG.debug("Processing event [{}]", event);
-
+        // fetch company
         String companyId = event.getPayload().getAccount().getAccountIdentifier();
+        Company company = companyDAO.findById(companyId);
+        if (company == null) {
+            throw new AccountNotFoundException(String.format("Could not find account [%s]", companyId));
+        }
         
         // delete company and users
-        Company company = companyDAO.findById(companyId);
         companyDAO.delete(company);
         
         return Result.successResult(String.format("Cancelled subscription for company [%s]", companyId));
